# REGR/PERF: Index.is_ performance regression